Refactors SmallTileRail into its own file.
