+++ CHANGELOG.yml
@@ -18,6 +18,7 @@ upcoming:
     - Adds other collections category collection hub rail - dzucconi
     - Fix for infinite spinner on inactive fairs - brian
     - Suggested artists now include the basis of that suggestion - ash
+    - Updated artwork tile design - ash & brian
 
 releases:
   - version: 6.4.3
